+++ changelog.md
@@ -2,6 +2,7 @@
 - The list ids command is now paginated
 - Add several bindings for noises
 - Add some additional helper functions for TFC's `Noise2D` and `Noise3D` classes
+- Add noise inspection commands
 
 ### 1.3.2
 - Add `#hasPostProcess` and `#emissiveRendering` methods to extended properties